test-suite/requirements.txt
pytest
python-dotenv
requests
streamlit
assemblyai
resend
pytube
moviepy
langchain_openai
twilio